requirements.txt
tldextract>=5.1.2
idna>=3.7
# Optional extras for live checks (enable as needed)
python-whois>=0.9.4
dnspython>=2.6.1
requests>=2.32.3
cryptography>=43.0.0
